pyproject.toml
[tool.poetry]
name = "dmql-database-project"
version = "0.1.0"
description = ""
authors = ["Amardhruva N Prabhu <amaramardhruva@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dmql_database_project"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.0"
psycopg2-binary = "^2.9.4"
ipykernel = "^6.16.0"
tqdm = "^4.64.1"
django = "^4.1.3"
django-crispy-forms = "^1.14.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
